+++ lib/ui/academics/EssentialSkillsCoachDashboardPanel.dart
@@ -1,4 +1,5 @@
 
+import 'package:collection/collection.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/model/CustomCourses.dart';
@@ -10,6 +11,7 @@ import 'package:illinois/ui/academics/EssentialSkillsCoach.dart';
 import 'package:illinois/ui/academics/courses/AssignmentPanel.dart';
 import 'package:illinois/ui/academics/courses/AssignmentCompletePanel.dart';
 import 'package:illinois/ui/academics/courses/ResourcesPanel.dart';
+import 'package:illinois/ui/academics/courses/SkillsHistoryPanel.dart';
 import 'package:illinois/ui/academics/courses/StreakPanel.dart';
 import 'package:illinois/ui/academics/courses/UnitInfoPanel.dart';
 import 'package:illinois/utils/AppUtils.dart';
@@ -20,6 +22,7 @@ import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 import 'package:sprintf/sprintf.dart';
 
+enum EssentialSkillsCoachTab { coach, history }
 
 class EssentialSkillsCoachDashboardPanel extends StatefulWidget {
   EssentialSkillsCoachDashboardPanel();
@@ -34,9 +37,9 @@ class _EssentialSkillsCoachDashboardPanelState extends State<EssentialSkillsCoac
   List<UserUnit>? _userCourseUnits;
   CourseConfig? _courseConfig;
   bool _loading = false;
-
   String? _selectedModuleKey;
   DateTime? _pausedDateTime;
+  EssentialSkillsCoachTab _selectedTab = EssentialSkillsCoachTab.coach;
 
   @override
   void initState() {
@@ -47,7 +50,6 @@ class _EssentialSkillsCoachDashboardPanelState extends State<EssentialSkillsCoac
     _loadCourseAndUnits();
     _loadCourseConfig();
     //TODO: check ESC onboarding completed, _hasStartedSkillsCoach, completed BESSI for onboarding sequence
-
     super.initState();
   }
 
@@ -65,17 +67,10 @@ class _EssentialSkillsCoachDashboardPanelState extends State<EssentialSkillsCoac
       if (_selectedModule != null) {
         return Column(
           children: [
-            _buildStreakWidget(),
-            Container(
-              color: _selectedModulePrimaryColor,
-              child: _buildModuleSelection(),
-            ),
+            _buildTabNavBar(),
             Expanded(
-              child: SingleChildScrollView(
-                child: Container(
-                  color: Styles().colors.background,
-                  child: Column(children: _buildModuleUnitWidgets(),),
-                ),
+              child: SizedBox.expand(
+                child: content
               ),
             ),
           ],
@@ -97,7 +92,18 @@ class _EssentialSkillsCoachDashboardPanelState extends State<EssentialSkillsCoac
 
   bool get _hasStartedSkillsCoach => _userCourse != null;
 
-  Module? get _selectedModule => _userCourse?.course?.searchByKey(moduleKey: _selectedModuleKey) ?? _course?.searchByKey(moduleKey: _selectedModuleKey); //TODO: remove _course option after start course UI
+  String? get _availableSelectedModuleKey {
+    List<DropdownMenuItem<String>> items = _moduleDropdownItems();
+    String? selected = items.firstOrNull?.value;
+    if(items.firstWhereOrNull((e) => e.value == _selectedModuleKey) != null) {
+      selected = _selectedModuleKey;
+    }
+    return selected;
+  }
+  Module? get _selectedModule {
+    String? selected = _availableSelectedModuleKey;
+    return _userCourse?.course?.searchByKey(moduleKey: selected) ?? _course?.searchByKey(moduleKey: selected);
+  } //TODO: remove _course option after start course UI
   Color? get _selectedModulePrimaryColor => _selectedModule!.styles?.colors?['primary'] != null ? Styles().colors.getColor(_selectedModule!.styles!.colors!['primary']!) : Styles().colors.fillColorPrimary;
   Color? get _selectedModuleAccentColor => _selectedModule!.styles?.colors?['accent'] != null ? Styles().colors.getColor(_selectedModule!.styles!.colors!['accent']!) : Styles().colors.fillColorSecondary;
   Widget? get _selectedModuleIcon => Styles().images.getImage(_selectedModule!.styles?.images?['icon'] ?? 'skills-question', size: 48);
@@ -107,6 +113,84 @@ class _EssentialSkillsCoachDashboardPanelState extends State<EssentialSkillsCoac
     if (name == AppLivecycle.notifyStateChanged) {
       _onAppLivecycleStateChanged(param);
     }
+  }
+
+  Widget get content {
+    if (_selectedTab == EssentialSkillsCoachTab.coach) {
+      return coachContent;
+    }
+    else if (_selectedTab == EssentialSkillsCoachTab.history) {
+      return historyContent;
+    }
+    return Container();
+  }
+
+  Widget get coachContent {
+    return Column(
+      children: [
+        _buildStreakWidget(),
+        Container(
+          color: _selectedModulePrimaryColor,
+          child: _buildModuleSelection(),
+        ),
+        Expanded(
+          child: SingleChildScrollView(
+            child: Container(
+              color: Styles().colors.background,
+              child: Column(children: _buildModuleUnitWidgets(),),
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+
+  Widget get historyContent {
+    return SkillsHistoryPanel();
+  }
+
+  String _tabLabel(EssentialSkillsCoachTab tab) {
+    if (tab == EssentialSkillsCoachTab.coach) {
+      return Localization().getStringEx('', "Skills Coach");
+    }
+    else if (tab == EssentialSkillsCoachTab.history) {
+      return Localization().getStringEx('', "Skills History");
+    }
+    return '';
+  }
+
+  Widget _buildTabButton(EssentialSkillsCoachTab tab) {
+    bool selected = _selectedTab == tab;
+    BorderRadius radius = BorderRadius.circular(40);
+    return Padding(
+      padding: const EdgeInsets.only(left: 8.0),
+      child: Material(
+        borderRadius: radius, // Creates border
+        color: selected ? Styles().colors.fillColorPrimary : Styles().colors.backgroundVariant,
+        child: InkWell(
+          borderRadius: radius,
+          onTap: () =>setStateIfMounted(() {
+            _selectedTab = tab;
+          }),
+          child: Padding(
+            padding: const EdgeInsets.symmetric(vertical: 8.0, horizontal: 16.0),
+            child: Text(_tabLabel(tab), style: selected ?
+              Styles().textStyles.getTextStyle("widget.title.light.small.fat")
+                : Styles().textStyles.getTextStyle("widget.title.small.fat")),
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildTabNavBar(){
+    return Padding(
+      padding: const EdgeInsets.all(8.0),
+      child: Row(children: [
+        _buildTabButton(EssentialSkillsCoachTab.coach),
+        _buildTabButton(EssentialSkillsCoachTab.history),
+      ]),
+    );
   }
 
   Widget _buildStreakWidget() {
@@ -144,6 +228,11 @@ class _EssentialSkillsCoachDashboardPanelState extends State<EssentialSkillsCoac
   }
 
   Widget _buildModuleSelection() {
+    List<DropdownMenuItem<String>> items = _moduleDropdownItems();
+    String? selected = items.firstOrNull?.value;
+    if(items.firstWhereOrNull((e) => e.value == _selectedModuleKey) != null) {
+      selected = _selectedModuleKey;
+    }
     return Padding(
       padding: const EdgeInsets.symmetric(vertical: 12.0),
       child: Row(
@@ -169,7 +258,7 @@ class _EssentialSkillsCoachDashboardPanelState extends State<EssentialSkillsCoac
                   alignedDropdown: true,
                   child: DropdownButton(
                       alignment: AlignmentDirectional.centerStart,
-                      value: _selectedModuleKey,
+                      value: selected,
                       iconDisabledColor: Styles().colors.fillColorSecondary,
                       iconEnabledColor: Styles().colors.fillColorSecondary,
                       focusColor: Styles().colors.surface,
@@ -177,7 +266,7 @@ class _EssentialSkillsCoachDashboardPanelState extends State<EssentialSkillsCoac
                       underline: Divider(color: Styles().colors.fillColorSecondary, height: 1.0, indent: 16.0, endIndent: 16.0),
                       borderRadius: BorderRadius.all(Radius.circular(4.0)),
                       isExpanded: true,
-                      items: _moduleDropdownItems(),
+                      items: items,
                       onChanged: (String? selected) {
                         setState(() {
                           _selectedModuleKey = selected;
@@ -453,43 +542,49 @@ class _EssentialSkillsCoachDashboardPanelState extends State<EssentialSkillsCoac
   }
 
   Future<void> _loadCourseAndUnits() async {
-    _userCourse ??= CustomCourses().userCourses?[Config().essentialSkillsCoachKey];
-    if (_userCourse == null) {
-      if (StringUtils.isNotEmpty(Config().essentialSkillsCoachKey)) {
-        _setLoading(true);
-        UserCourse? userCourse = await CustomCourses().loadUserCourse(Config().essentialSkillsCoachKey!);
-        if (userCourse != null) {
-          setStateIfMounted(() {
-            _userCourse = userCourse;
-            _selectedModuleKey ??= CollectionUtils.isNotEmpty(userCourse.course?.modules) ? userCourse.course!.modules![0].key : null;
-            _loading = false;
+    if (mounted) {
+      _userCourse ??= CustomCourses().userCourses?[Config().essentialSkillsCoachKey];
+      if (_userCourse == null) {
+        if (StringUtils.isNotEmpty(Config().essentialSkillsCoachKey)) {
+          setState(() {
+            _loading = true;
           });
-          await _loadUserCourseUnits();
-        } else {
-          await _loadCourse();
+          UserCourse? userCourse = await CustomCourses().loadUserCourse(Config().essentialSkillsCoachKey!);
+          if (mounted) {
+            if (userCourse != null) {
+              setState(() {
+                _userCourse = userCourse;
+                _selectedModuleKey ??= CollectionUtils.isNotEmpty(userCourse.course?.modules) ? userCourse.course!.modules![0].key : null;
+                _loading = false;
+              });
+              await _loadUserCourseUnits();
+            } else {
+              await _loadCourse();
+            }
+          }
         }
+      } else {
+        _selectedModuleKey ??= CollectionUtils.isNotEmpty(_userCourse?.course?.modules) ? _userCourse?.course!.modules![0].key : null;
+        await _loadUserCourseUnits();
       }
-    } else {
-      _selectedModuleKey ??= CollectionUtils.isNotEmpty(_userCourse?.course?.modules) ? _userCourse?.course!.modules![0].key : null;
-      await _loadUserCourseUnits();
     }
   }
 
   Future<void> _loadCourse() async {
     _course ??= CustomCourses().courses?[Config().essentialSkillsCoachKey];
     if (_course == null) {
-      if (StringUtils.isNotEmpty(Config().essentialSkillsCoachKey)) {
-        _setLoading(true);
+      if (StringUtils.isNotEmpty(Config().essentialSkillsCoachKey) && mounted) {
+        setState(() {
+          _loading = true;
+        });
         Course? course = await CustomCourses().loadCourse(Config().essentialSkillsCoachKey!);
-        if (course != null) {
-          setStateIfMounted(() {
+        setStateIfMounted(() {
+          if (course != null) {
             _course = course;
             _selectedModuleKey ??= CollectionUtils.isNotEmpty(course.modules) ? course.modules![0].key : null;
-            _loading = false;
-          });
-        } else {
-          _setLoading(false);
-        }
+          }
+          _loading = false;
+        });
       }
     } else {
       setStateIfMounted(() {
@@ -501,104 +596,112 @@ class _EssentialSkillsCoachDashboardPanelState extends State<EssentialSkillsCoac
 
   Future<void> _loadUserCourseUnits() async {
     _userCourseUnits ??= CustomCourses().userCourseUnits?[Config().essentialSkillsCoachKey];
-    if (_userCourseUnits == null) {
-      if (StringUtils.isNotEmpty(Config().essentialSkillsCoachKey)) {
-        _setLoading(true);
-        List<UserUnit>? userUnits = await CustomCourses().loadUserCourseUnits(Config().essentialSkillsCoachKey!);
+    if ((_userCourseUnits == null) && StringUtils.isNotEmpty(Config().essentialSkillsCoachKey) && mounted) {
+      setState(() {
+        _loading = true;
+      });
+      List<UserUnit>? userUnits = await CustomCourses().loadUserCourseUnits(Config().essentialSkillsCoachKey!);
+      setStateIfMounted(() {
         if (userUnits != null) {
-          setStateIfMounted(() {
-            _userCourseUnits = userUnits;
-          });
+          _userCourseUnits = userUnits;
         }
-      }
+        _loading = false;
+      });
     }
-    _setLoading(false);
   }
 
   Future<void> _loadCourseConfig() async {
     if (_courseConfig == null && StringUtils.isNotEmpty(Config().essentialSkillsCoachKey)) {
-      _setLoading(true);
+      setState(() {
+        _loading = true;
+      });
       CourseConfig? courseConfig = await CustomCourses().loadCourseConfig(Config().essentialSkillsCoachKey!);
-      if (courseConfig != null) {
-        setStateIfMounted(() {
+      setStateIfMounted(() {
+        if (courseConfig != null) {
           _courseConfig = courseConfig;
-          _loading = false;
-        });
-      } else {
-        _setLoading(false);
-      }
+        }
+        _loading = false;
+      });
     }
   }
 
-  Future<void> _startCourse() async {
-    await _loadCourseAndUnits();
-    if (_userCourse == null && StringUtils.isNotEmpty(Config().essentialSkillsCoachKey)) {
-      _setLoading(true);
+  Future<void> _startCourse(String? moduleKeyPrefix) async {
+    //await _loadCourseAndUnits();
+    if (_userCourse == null && StringUtils.isNotEmpty(Config().essentialSkillsCoachKey) && mounted) {
+      setState(() {
+        _loading = true;
+      });
       UserCourse? userCourse = await CustomCourses().createUserCourse(Config().essentialSkillsCoachKey!);
-      if (userCourse != null) {
-        setStateIfMounted(() {
+      setStateIfMounted(() {
+        if (userCourse != null) {
           _userCourse = userCourse;
-        });
-      }
+          String moduleKey = "${moduleKeyPrefix}_skills";
+          if(StringUtils.isNotEmpty(moduleKeyPrefix)) {
+            _selectedModuleKey = moduleKey;
+          }
+        }
+        _loading = false;
+      });
     }
-    _setLoading(false);
   }
 
   Future<void> _updateProgress(String moduleKey, String unitKey, Map<String, dynamic> response, UserContentReference userContentReference, int unitNumber, int activityNumber) async {
-    _setLoading(true);
-    UserUnit? updatedUserUnit = await CustomCourses().updateUserCourseProgress(UserResponse(unitKey: unitKey, contentKey: userContentReference.contentKey!, response: response), courseKey: _userCourse!.course!.key!, moduleKey: moduleKey);
-    if (updatedUserUnit != null) {
-      if (CollectionUtils.isNotEmpty(_userCourseUnits)) {
-        int unitIndex = _userCourseUnits!.indexWhere((userUnit) => userUnit.id != null && userUnit.id == updatedUserUnit.id);
-        if (unitIndex >= 0) {
-          setStateIfMounted(() {
-            _userCourseUnits![unitIndex] = updatedUserUnit;
-          });
-        } else {
-          setStateIfMounted(() {
-            _userCourseUnits!.add(updatedUserUnit);
-          });
-        }
-      } else {
-        setStateIfMounted(() {
-          _userCourseUnits ??= [];
-          _userCourseUnits!.add(updatedUserUnit);
-        });
-      }
+    if (mounted) {
+      setState(() {
+        _loading = true;
+      });
+      UserUnit? updatedUserUnit = await CustomCourses().updateUserCourseProgress(UserResponse(unitKey: unitKey, contentKey: userContentReference.contentKey!, response: response), courseKey: _userCourse!.course!.key!, moduleKey: moduleKey);
+      if (mounted) {
+        if (updatedUserUnit != null) {
+          if (CollectionUtils.isNotEmpty(_userCourseUnits)) {
+            int unitIndex = _userCourseUnits!.indexWhere((userUnit) => userUnit.id != null && userUnit.id == updatedUserUnit.id);
+            if (unitIndex >= 0) {
+              setStateIfMounted(() {
+                _userCourseUnits![unitIndex] = updatedUserUnit;
+              });
+            } else {
+              setStateIfMounted(() {
+                _userCourseUnits!.add(updatedUserUnit);
+              });
+            }
+          } else {
+            setStateIfMounted(() {
+              _userCourseUnits ??= [];
+              _userCourseUnits!.add(updatedUserUnit);
+            });
+          }
 
-      bool earnedPause = false;
-      bool extendedStreak = false;
-      UserCourse? userCourse = await CustomCourses().loadUserCourse(Config().essentialSkillsCoachKey!);
-      if (userCourse != null) {
-        earnedPause = (userCourse.pauses ?? 0) > (_userCourse?.pauses ?? 0);
-        extendedStreak = (userCourse.streak ?? 0) > (_userCourse?.streak ?? 0);
-        setStateIfMounted(() {
-          _userCourse = userCourse;
-          _loading = false;
-        });
-      } else {
-        _setLoading(false);
-      }
+          bool earnedPause = false;
+          bool extendedStreak = false;
+          UserCourse? userCourse = await CustomCourses().loadUserCourse(Config().essentialSkillsCoachKey!);
+          if (mounted) {
+            setState(() {
+              if (userCourse != null) {
+                earnedPause = (userCourse.pauses ?? 0) > (_userCourse?.pauses ?? 0);
+                extendedStreak = (userCourse.streak ?? 0) > (_userCourse?.streak ?? 0);
+                _userCourse = userCourse;
+              }
+              _loading = false;
+            });
 
       // if the current task was just completed and it extended the user's streak
       if (extendedStreak && userContentReference.isNotComplete && response[UserContent.completeKey] == true) {
-        Navigator.push(context, CupertinoPageRoute(builder: (context) => AssignmentCompletePanel(
-          unitNumber: unitNumber,
-          activityNumber: activityNumber,
-          pauses: earnedPause ? _userCourse?.pauses : null,
-          color: _selectedModulePrimaryColor,
-        )));
+        Navigator.push(context, CupertinoPageRoute(builder: (context) =>
+            AssignmentCompletePanel(
+              unitNumber: unitNumber,
+              activityNumber: activityNumber,
+              pauses: earnedPause ? _userCourse?.pauses : null,
+              color: _selectedModulePrimaryColor,
+            )));
+        }
       }
     } else {
-      _setLoading(false);
+      setState(() {
+        _loading = false;
+      });;
     }
   }
 
-  void _setLoading(bool value) {
-    setStateIfMounted(() {
-      _loading = value;
-    });
-  }
 
   void _onAppLivecycleStateChanged(AppLifecycleState? state) {
     if (state == AppLifecycleState.paused) {