+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [5.1.26] - 2024-02-12
 ### Added
+- Essential Skills History [#3964](https://github.com/rokwire/illinois-app/issues/3964)
+- Essential Skills Coach Onboarding Flow [#3950](https://github.com/rokwire/illinois-app/issues/3950)
 - Essential Skills Coach assignment response history [#3966](https://github.com/rokwire/illinois-app/issues/3966).
+### Fixed
+- Do not allow manual attendance taking if relevant options are not selected. 
 
 ## [5.1.25] - 2024-02-09
 ### Changed